refactor: move subagent tracking from hook_events into claude_task_state
Relocate _active_subagents dict, add_subagent(), remove_subagent(),
get_subagent_names(), build_subagent_label(), and clear_subagents()
into claude_task_state.py, which already owns per-window task snapshots.
hook_events.py now imports and calls these functions instead of maintaining
its own parallel data structure.
